+++ CHANGELOG.md
@@ -6,6 +6,8 @@ project adheres to [Semantic Versioning][SemVer].
 ## [Unreleased]
 ### Changed
 - Use a (pseudo)random `secret` by default.
+- Improve autodetection of relay IP addresses used by default if the
+  `relay_ipv4_addr` and/or `relay_ipv6_addr` options aren't specified.
 
 ### Fixed
 - Don't crash without explicit `listen` configuration. This bug was introduced